{{cookiecutter.project_name}}/requirements.txt
bottle
psycopg2
pytest
bottle-cerberus
pyjwt
jinja2
gunicorn
meinheld
git+http://github.com/rgercp/pydbwrapper.git#egg=pydbwrapper
